requirements.txt
azure-core==1.33.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
